- added common money struct - use money struct instead of fixed price
