+++ src/PayPal/Api/Plan/BillingCycle/PricingScheme/FixedPrice.php
@@ -7,37 +7,8 @@
 
 namespace Swag\PayPal\PayPal\Api\Plan\BillingCycle\PricingScheme;
 
-use Swag\PayPal\PayPal\Api\Common\PayPalStruct;
+use Swag\PayPal\PayPal\Api\Common\Money;
 
-class FixedPrice extends PayPalStruct
+class FixedPrice extends Money
 {
-    /**
-     * @var string
-     */
-    protected $value;
-
-    /**
-     * @var string
-     */
-    protected $currencyCode;
-
-    public function getValue(): string
-    {
-        return $this->value;
-    }
-
-    public function setValue(string $value): void
-    {
-        $this->value = $value;
-    }
-
-    public function getCurrencyCode(): string
-    {
-        return $this->currencyCode;
-    }
-
-    public function setCurrencyCode(string $currencyCode): void
-    {
-        $this->currencyCode = $currencyCode;
-    }
 }
